+++ packages/block-library/src/accordion-panel/icon.js
@@ -3,7 +3,7 @@
  */
 import { SVG, Path } from '@wordpress/primitives';
 
-const accordionPanel = (
+export default (
 	<SVG xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
 		<Path
 			fillRule="evenodd"
@@ -13,5 +13,3 @@ const accordionPanel = (
 		/>
 	</SVG>
 );
-
-export default accordionPanel;