+++ packages/block-library/src/accordions/index.js
@@ -1,21 +1,18 @@
 /**
- * WordPress dependencies
- */
-import { accordion } from '@wordpress/icons';
-/**
  * Internal dependencies
  */
 import edit from './edit';
 import save from './save';
 import metadata from './block.json';
 import initBlock from '../utils/init-block';
+import icon from './icon';
 
 const { name } = metadata;
 
 export { metadata, name };
 
 export const settings = {
-	icon: accordion,
+	icon,
 	example: {},
 	edit,
 	save,